backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.4.0
Pillow>=9.0.0
